Add js docs for all new files, reorganize code a bit into sub-folders
