+++ src/ai/llm/factory.ts
@@ -5,10 +5,10 @@ import { logger } from '../../utils/logger.js';
 import { openai } from '@ai-sdk/openai';
 import { google } from '@ai-sdk/google';
 import { anthropic } from '@ai-sdk/anthropic';
-import { VercelLLMService } from './vercel.js';
+import { VercelLLMService } from './services/vercel.js';
 import { VercelLLM } from './types.js';
-import { OpenAIService } from './openai.js';
-import { AnthropicService } from './anthropic.js';
+import { OpenAIService } from './services/openai.js';
+import { AnthropicService } from './services/anthropic.js';
 /**
  * Extract and validate API key from config or environment variables
  * @param config LLM configuration from the config file
