+++ src/ai/llm/message/formatters/vercel.ts
@@ -1,7 +1,25 @@
 import { IMessageFormatter } from '../formatter.js';
 import { InternalMessage } from '../types.js';
 
+/**
+ * Message formatter for Vercel AI SDK.
+ * 
+ * Converts the internal message format to Vercel's specific structure:
+ * - System prompt is included in the messages array
+ * - Tool calls use function_call property instead of tool_calls
+ * - Tool results use the 'function' role instead of 'tool'
+ * 
+ * Note: Vercel's implementation is different from OpenAI's standard,
+ * particularly in its handling of function calls and responses.
+ */
 export class VercelFormatter implements IMessageFormatter {
+    /**
+     * Formats internal messages into Vercel AI SDK format
+     * 
+     * @param history Array of internal messages to format
+     * @param systemPrompt System prompt to include at the beginning of messages
+     * @returns Array of messages formatted for Vercel's API
+     */
     format(history: Readonly<InternalMessage[]>, systemPrompt: string | null): any[] {
         const formatted = [];
 
@@ -56,7 +74,13 @@ export class VercelFormatter implements IMessageFormatter {
         return formatted;
     }
 
-    // Vercel includes system prompt in the messages array
+    /**
+     * Vercel handles system prompts in the messages array
+     * This method returns null since the system prompt is already
+     * included directly in the formatted messages.
+     * 
+     * @returns null as Vercel doesn't need a separate system prompt
+     */
     getSystemPrompt(): null {
         return null;
     }
