+++ src/ai/llm/services/anthropic.ts
@@ -1,11 +1,11 @@
 import Anthropic from '@anthropic-ai/sdk';
-import { ClientManager } from '../../client/manager.js';
-import { ILLMService } from './types.js';
-import { ToolSet } from '../types.js';
-import { logger } from '../../utils/logger.js';
+import { ClientManager } from '../../../client/manager.js';
+import { ILLMService } from '../types.js';
+import { ToolSet } from '../../types.js';
+import { logger } from '../../../utils/logger.js';
 import { EventEmitter } from 'events';
-import { MessageManager } from './message/manager.js';
-import { AnthropicFormatter } from './message/formatters/anthropic.js';
+import { MessageManager } from '../message/manager.js';
+import { AnthropicFormatter } from '../message/formatters/anthropic.js';
 
 /**
  * Anthropic implementation of LLMService
