+++ src/ai/llm/services/openai.ts
@@ -1,11 +1,11 @@
 import OpenAI from 'openai';
-import { ClientManager } from '../../client/manager.js';
-import { ILLMService } from './types.js';
-import { ToolSet } from '../types.js';
-import { logger } from '../../utils/logger.js';
+import { ClientManager } from '../../../client/manager.js';
+import { ILLMService } from '../types.js';
+import { ToolSet } from '../../types.js';
+import { logger } from '../../../utils/logger.js';
 import { EventEmitter } from 'events';
-import { MessageManager } from './message/manager.js';
-import { OpenAIFormatter } from './message/formatters/openai.js';
+import { MessageManager } from '../message/manager.js';
+import { OpenAIFormatter } from '../message/formatters/openai.js';
 
 // System prompt constants
